refactor: move API to src
